Don't use "./node_modules/" in the import path
When using webpack, modules from dependencies are usually not included through their actual paths on the file system. Instead, the include path starts with the name of the module and it's webpack's job to find the correct directory.

On the other hand, having the full path here is nice because it tells the reader where to go looking for the actual files that wasm-pack produced. So I'm not sure if this is a good change.